Factor out job-coordinator job summary
